+++ domain/src/main/java/joist/domain/orm/queries/Select.java
@@ -114,8 +114,8 @@ public class Select<T extends DomainObject> {
     return results.get(0);
   }
 
-  public List<Integer> listIds() {
-    List<Integer> ids = new ArrayList<Integer>();
+  public List<Long> listIds() {
+    List<Long> ids = new ArrayList<Long>();
     this.select(this.from.getIdColumn().as("id"));
     this.orderBy(this.from.getIdColumn().asc()); // determinism
     Jdbc.query(UoW.getConnection(), this.toSql(), this.getParameters(), new IdsMapper<T>(this.from, ids));
